Remove re normalizer
