+++ plugin/action/hash/hash.go
@@ -114,8 +114,6 @@ type Config struct {
 	// > The event field to which put the hash.
 	HashField  cfg.FieldSelector `json:"hash_field" parse:"selector" required:"true"` // *
 	HashField_ []string
-
-	Normalizer string `json:"normalizer" default:"re" options:"re|lm"`
 }
 
 type fieldFormat byte
@@ -147,12 +145,14 @@ func factory() (pipeline.AnyPlugin, pipeline.AnyConfig) {
 func (p *Plugin) Start(config pipeline.AnyConfig, params *pipeline.ActionPluginParams) {
 	p.config = config.(*Config)
 
-	if p.config.Normalizer == "re" {
-		p.normalizer = normalize.NewReNormalizer()
-	} else {
-		p.normalizer = normalize.NewLMNormalizer()
-	}
 	p.buf = make([]byte, 0)
+
+	for _, f := range p.config.Fields {
+		if f.Format_ == ffNormalize {
+			p.normalizer = normalize.NewTokenNormalizer()
+			break
+		}
+	}
 }
 
 func (p *Plugin) Stop() {}
