+++ plugin/action/hash/normalize/normalizer.go
@@ -4,7 +4,7 @@ type Normalizer interface {
 	Normalize(out, data []byte) []byte
 }
 
-func formatPlaceholder(out []byte, name string) []byte {
+func addPlaceholder(out []byte, name string) []byte {
 	out = append(out, '<')
 	out = append(out, name...)
 	out = append(out, '>')
